Fix bug where uninitialized char arrays will cause the client to serialize more data than it should
This is due to the fact that all char arrays, regardless of initialization status, are serialized based on the assumption that they are null-terminated. However, since the contents of uninitialized char arrays are unspecified, it is possible that a NULL terminator shows up thousand characters later.

This was clearly a problem: for most char arrays, the host has a preallocated buffer of a specific length, and any write past that buffer can lead to all sorts of issues.

This commit fixes that by constraining the amount of data set to the maximum limit specified in the Vulkan registry.