Remove code in codegen that effectively skipped aliases in vkGet*ProcAddr
This was needed in the past, as the driver didn't handle aliases, so it was important to force the loader to look for the "actual" command

However, recently, for at least one alias (vkGetPhysicalDeviceCalibrateableTimeDomainsEXT), if it can not find it, the loader does not fall back. So, by returning success for supported aliases, this failure can be avoided.